Implement open hours in menu

+++ app/src/main/java/org/cornelldti/flux/data/Facility.kt
@@ -91,8 +91,8 @@ data class MenuCategory(
 data class Menu(
     val menu: List<MenuCategory>,
     val description: Meal,
-    val endTime: Int,
-    val startTime: Int,
+    val endTime: Long,
+    val startTime: Long,
     // val similarity: Double
 )
